NR-84070 - Leverage commit SHA collected by agent to handle Node anon functions - 🔥highly demoable - Fix tests
TODO
No commit sha scenario
Possibly recommend get commit sha
Caching / batching
Dark theme look and feel
Clickly to get graphs

+++ jb/src/main/kotlin/protocols/agent/AgentProtocol.kt
@@ -229,10 +229,10 @@ data class Markerish(val id: String, val referenceLocations: Array<CSReferenceLo
 
 data class CSMarkerLocation(
     val id: String,
-    val lineStart: Long,
-    val colStart: Long,
-    val lineEnd: Long,
-    val colEnd: Long,
+    val lineStart: Int,
+    val colStart: Int,
+    val lineEnd: Int,
+    val colEnd: Int,
     val meta: CSLocationMeta?,
 )
